+++ packages/apollo-server/src/index.ts
@@ -1,5 +1,6 @@
 import * as express from 'express';
 import { registerServer } from 'apollo-server-express';
+import { OptionsJson } from 'body-parser';
 
 import {
   ApolloServerBase,
@@ -11,38 +12,31 @@ import {
 export * from './exports';
 
 export class ApolloServer extends ApolloServerBase<express.Request> {
-  private disableHealthCheck: boolean = false;
-  private onHealthCheck: (req: express.Request) => Promise<any>;
-
-  constructor({
-    disableHealthCheck,
-    onHealthCheck,
-    ...opts
-  }: Config<express.Request> & {
-    onHealthCheck?: (req: express.Request) => Promise<any>;
-    disableHealthCheck?: boolean;
-  }) {
-    super(opts);
-    if (disableHealthCheck) this.disableHealthCheck = true;
-    this.onHealthCheck = onHealthCheck;
-  }
-
   // here we overwrite the underlying listen to configure
   // the fallback / default server implementation
-  async listen(opts: ListenOptions = {}): Promise<ServerInfo> {
+  async listen(
+    opts: ListenOptions & {
+      onHealthCheck?: (req: express.Request) => Promise<any>;
+      disableHealthCheck?: boolean;
+      bodyParserConfig?: OptionsJson;
+    } = {},
+  ): Promise<ServerInfo> {
+    //defensive copy
+    const { onHealthCheck } = opts;
+
     // we haven't configured a server yet so lets build the default one
     // using express
     if (!this.getHttp) {
       const app = express();
 
-      if (!this.disableHealthCheck) {
+      if (!opts.disableHealthCheck) {
         //uses same path as engine
         app.use('/.well-known/apollo/server-health', (req, res, next) => {
           //Response follows https://tools.ietf.org/html/draft-inadarei-api-health-check-01
           res.type('application/health+json');
 
-          if (this.onHealthCheck) {
-            this.onHealthCheck(req)
+          if (onHealthCheck) {
+            onHealthCheck(req)
               .then(() => {
                 res.json({ status: 'pass' });
               })
@@ -53,9 +47,9 @@ export class ApolloServer extends ApolloServerBase<express.Request> {
             res.json({ status: 'pass' });
           }
         });
-
-        await registerServer({ app, path: '/', server: this });
       }
+
+      await registerServer({ app, path: '/', server: this });
     }
 
     return super.listen(opts);